Move loader utils into utils
